requirements.txt
openai>=1.30.0
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
docker>=7.0.0
PyGithub>=2.3.0
click>=8.1.0
colorlog>=6.8.0
